revise footer
