+++ packages/app/src/Router.tsx
@@ -11,8 +11,8 @@ import { useAccountFromUrl } from 'hooks/useAccountFromUrl'
 import { ErrorFallbackApp, ErrorFallbackRoutes } from 'library/ErrorBoundary'
 import { Headers } from 'library/Headers'
 import { Help } from 'library/Help'
+import { MainFooter } from 'library/MainFooter'
 import { Menu } from 'library/Menu'
-import { NetworkBar } from 'library/NetworkBar'
 import { NotificationPrompts } from 'library/NotificationPrompts'
 import { Offline } from 'library/Offline'
 import { PageWithTitle } from 'library/PageWithTitle'
@@ -89,9 +89,9 @@ const RouterInner = () => {
               />
             </Routes>
           </ErrorBoundary>
+          <MainFooter />
         </Main>
       </Body>
-      <NetworkBar />
       <Offline />
     </ErrorBoundary>
   )
