+++ packages/styles/theme/theme.scss
@@ -27,7 +27,6 @@ SPDX-License-Identifier: GPL-3.0-only */
   --background-canvas-card: rgb(245 245 245 / 90%);
   --background-canvas-card-secondary: rgb(255 255 255 / 35%);
   --background-floating-card: rgb(255 255 255 / 90%);
-  --background-app-footer: rgb(244 225 225 / 75%);
   --background-modal: #f9f7f7;
   --background-modal-footer: #efefef;
   --background-status-overlay: rgb(255 255 255 / 85%);
@@ -105,7 +104,6 @@ SPDX-License-Identifier: GPL-3.0-only */
   --background-canvas-card: rgb(44 40 44 / 90%);
   --background-canvas-card-secondary: rgb(44 40 44 / 35%);
   --background-floating-card: rgb(43 38 43 / 95%);
-  --background-app-footer: #262327;
   --background-modal: rgb(43 38 43);
   --background-modal-footer: rgb(37 32 37);
   --background-status-overlay: rgb(43 38 43 / 75%);
