+++ packages/ui-structure/src/Footer/index.module.scss
@@ -1,10 +1,9 @@
 /* Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-.page {
+.footer {
   display: flex;
   flex-direction: column;
-  padding-bottom: 0.75rem;
   width: 100%;
   margin: 0 auto;
 }
